services/ingest_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
gunicorn==21.2.0
redis==5.0.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
pydantic==2.5.2
